Update after refactoring because of change of project name
