+++ app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.luan.hsworms.hydroit"
+        applicationId "com.luan.hsworms.hydroid"
         minSdkVersion 21
         targetSdkVersion 30
         versionCode 1
