+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.luan.hsworms.hydroit">
+    package="com.luan.hsworms.hydroid">
 
     <application
         android:allowBackup="true"
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Hydroit">
-        <activity android:name=".MainActivity">
+        <activity android:name="com.luan.hsworms.hydroid.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
