+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/logoFragment"
-        android:name="com.luan.hsworms.hydroit.LogoFragment"
+        android:name="com.luan.hsworms.hydroid.LogoFragment"
         android:label="fragment_logo"
         tools:layout="@layout/fragment_logo" >
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/userInputFragment"
-        android:name="com.luan.hsworms.hydroit.UserInputFragment"
+        android:name="com.luan.hsworms.hydroid.UserInputFragment"
         android:label="user_input_fragment"
         tools:layout="@layout/user_input_fragment" />
 </navigation>
