+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">Hydroit</string>
-    <string name="logo_hydroit">Hydroit</string>
+    <string name="app_name">Hydroid</string>
+    <string name="logo_hydroit">Hydroid</string>
     <string name="logo_fragment_background_image">Glass of Water</string>
 </resources>
